# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inosoft Test",
        "description": "Inosoft Test Docs API",
        "contact": {
            "email": "anggawijaya4567@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/sales/report": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "description": "Report Sales Endpoint",
                "operationId": "0109b551eb6084faee49b22313dc55c0",
                "parameters": [
                    {
                        "name": "vehicle_type",
                        "in": "query",
                        "description": "Vehicle Type. Ex: motorcycle|car",
                        "required": false
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/sales": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "description": "Sales Create Endpoint",
                "operationId": "51817db78f9cc2d36dcef3a5346c12ef",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "blue"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "format": "test",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Register Endpoint",
                "operationId": "526ace942a6d422b3682a486dfc18d73",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "name": {
                                        "type": "boolean",
                                        "example": "angga"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Login Endpoint",
                "operationId": "d333c91f7961ca3b3afc478978d1be78",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/auth/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "description": "profile Endpoint",
                "operationId": "e90ee6e0140f6e4744f9c7f90d6dff09",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "description": "logout Endpoint",
                "operationId": "2e501a116aebee6cc5d646f9672ec1c9",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/vehicles": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "description": "Vehicle Index Endpoint",
                "operationId": "ec407c20094a18b143c36165ac4faf7d",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicle"
                ],
                "description": "Vehicle Create Endpoint",
                "operationId": "30aac71ccd53f75477209f9b585f03a5",
                "requestBody": {
                    "description": "Example: https://pastebin.com/Yd5aS7yy",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "description": "Vehicle Get Endpoint",
                "operationId": "59a5fab57558be22007c4dfd5e7614bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of vehicle",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicle"
                ],
                "description": "Vehicle Update Endpoint",
                "operationId": "f0065a4ff205273c1200e575db5f4785",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of vehicle",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Example: https://pastebin.com/Yd5aS7yy",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "color",
                                    "year",
                                    "price"
                                ],
                                "properties": {
                                    "color": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "blue"
                                    },
                                    "year": {
                                        "type": "string",
                                        "format": "test",
                                        "example": "1000"
                                    },
                                    "price": {
                                        "type": "string",
                                        "format": "test",
                                        "example": "1000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicle"
                ],
                "description": "Vehicle Delete Endpoint",
                "operationId": "97b7203f4766cb41afb4ae504fd56cdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of vehicle",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}